Rewrite `stackFst` and `stackSnd` in terms of `split`
It turns out that the choice of induction I used for `stackSnd` was overly
complex, which made the proofs harder, and it would have been less runtime-
efficient, too. I realized this when I decided to define `Split` as a view,
made a function `split` to construct that view, and realized I basically
implemented both `stackFst` and `stackSnd` at once. This speaks to the
power of types to help with program and proof construction.